refactor: reuse VariableSelect in verification editor
